Keep Showing A View of your hierarchy.
If you want to keep a certain view of you heriarchy after the CoachMarksController starts like, for example the navigationItem.titleView, you can use this keepShowingView.
This view takes two identical views:
	- originalView that will be used to determine what are the constraints that should be installed on the cloneView
	 - cloneView an exact copy of the originalView. It will be places at the same place of the originalView but will not be covered by the overlay.
The cloneView is used to avoid breaking and restoring back constraints of the original view.
Future enhancements are to send only the originalView and the system will try to save the already installed constraints and restore them when the guide tour was completed.